LTSView should now compile without any warnings.
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@1075 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ src/ltsview/visualizer.cpp
@@ -80,7 +80,7 @@ void Visualizer::positionClusters()
   {
     // iterate over the clusters in this rank
     vector< Cluster* > clusters;
-    lts->getClustersAtRank( rank, clusters );
+    lts->getClustersAtRank( (unsigned int) rank, clusters );
     vector< Cluster* >::iterator clusit;
     
     for ( clusit = clusters.begin() ; clusit != clusters.end() ; ++clusit )
